% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BRCA_met.R
\docType{data}
\name{BRCA_met}
\alias{BRCA_met}
\title{Methylation levels for different patients with BRCA cancer.}
\format{data.frame}
\description{
Rows of this data set correspond to patients, more precisely, to samples taken from patients.
First column \code{SUBTYPE} corresponds to a subtype of BRCA cancer, next columns correspond to CpG islands form 17th chromosome.
Values inside the table indicate the methylation level of CpG island for specified sample.
}
